move query/mutation into folders
